Resolve issues with Symfony dependencies
This commit removes the dependency on the Symfony Config component. The
only code other than getters that was actually being used was the
constructors. In symfony 2, these were incredibly basic. In symfony 3,
they perform some validation.

This validation was causing a couple of "does not exist" tests to fail,
as the Symfony Config module was performing these checks before the
Lurker module, and throwing a different exception. Removing this
dependency neatly solves this issue at the same time.