Disable InotifyTracker due to faulty rename handling. Update tracker tests.
This is a follow-up to the previous commits involving renamed subdirectories
(eg fe7e4244dd4d5191c1a14c62239d51ba3cb144ef).  I had thought it merely
under-reported some aspects of the renaming. However, if you add
a little more on to the test (e.g. rename a folder... and then
add a file to that folder), then it starts to generate hard errors.

As newcomer to the codebase, I find the model a little mind-bending, and I
don't see a quick fix.  So I'll just disable it.  File-polling is less
performant, but at least it works...  Maybe someone else can fix/rewrite the
`inotify` driver (or maybe someday I'll take a fresh look).